requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-google-genai>=0.0.4
google-generativeai>=0.3.2
python-dotenv>=0.19.0
chromadb>=0.4.22
streamlit>=1.31.0
markdown>=3.5.2
python-magic>=0.4.27
unstructured>=0.11.6
sentence-transformers>=2.2.2
torch>=2.2.0
transformers>=4.37.2